backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-beast"
version = "0.1.0"
description = "Multi-agent AI assistant platform with pluggable data adapters"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "TNN AI Project", email = "team@example.com"}
]
license = {text = "MIT"}

dependencies = [
    # FastAPI & Web
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",

    # Database & Async
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13.0",
    "pgvector>=0.2.0",

    # Data Validation & Settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.1.0",

    # Background Tasks & Scheduling
    "celery>=5.3.0",
    "redis>=5.0.0",
    "apscheduler[sqlalchemy]>=3.10.0",

    # Data Processing
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "PyPDF2>=3.0.0",

    # AI & LLMs
    "openai>=1.3.0",
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",
    "sentence-transformers>=2.2.0",
    "motor>=3.3.0",

    # Agent Framework
    "strands-agents[openai]>=1.37.0",
    "strands-agents[litellm]>=1.37.0",
    "strands-agents[ollama]>=1.37.0",
    "strands-agents-tools>=0.5.1",
    "ollama>=0.4.0",
    "langchain-text-splitters>=0.0.1",
    
    # tools 
    "sqlglot>=20.0",

    # Monitoring & Metrics
    "prometheus-client>=0.19.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "cryptography>=41.0.0",

    # Email & Gmail
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",

    # Observability & Logging
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "sentry-sdk[fastapi]>=1.40.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "sqlparse>=0.4.4",
    "PyYAML>=6.0",

    # Code Interpreter
    "RestrictedPython>=7.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
multimodal = [
    "huggingface_hub>=0.30.0",
    "litert-lm>=0.10.1",
    "numpy>=2.0.0",
    "mlx-audio>=0.4.2; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "kokoro-onnx>=0.5.0; sys_platform == 'linux'",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.26.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "ipython>=8.18.0",
    "ipdb>=0.13.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "slow: Slow tests",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
